remove legacy delete methods, update test cases
