+++ crates/content-base/src/db/op/mod.rs
@@ -1,3 +1,3 @@
 mod create;
 mod delete;
-mod select;
+mod test;
