+++ crates/content-base/src/db/search/test.rs
@@ -1,75 +1,7 @@
 #[cfg(test)]
 mod tests {
     use crate::db::shared::test::setup;
-    // use itertools::Itertools;
-    // use std::process::id;
     use test_log::test;
-
-    #[test(tokio::test)]
-    async fn test_select_text() {
-        let db = setup(None).await;
-        let text_res = db
-            .select_text(vec!["text:7dd12x11yvt5fgamdjb0"])
-            .await
-            .unwrap();
-        println!("text_res: {:?}", text_res);
-    }
-
-    #[test(tokio::test)]
-    async fn test_select_image() {
-        let db = setup(None).await;
-        let image_res = db
-            .select_image(vec!["image:flzkn6ncniglqttxnrsm"])
-            .await
-            .unwrap();
-        println!("image_res: {:?}", image_res);
-    }
-
-    #[test(tokio::test)]
-    async fn test_select_audio() {
-        let db = setup(None).await;
-        let audio_res = db
-            .select_audio(vec!["audio:gkzq6db9jwr34l3j0gmz"])
-            .await
-            .unwrap();
-        println!("audio_res: {:?}", audio_res);
-    }
-
-    #[test(tokio::test)]
-    async fn test_select_video() {
-        let db = setup(None).await;
-        let video_res = db
-            .select_video(vec!["video:u456grwuvl6w74zgqemc"])
-            .await
-            .unwrap();
-        println!("video_res: {:?}", video_res);
-    }
-
-    #[test(tokio::test)]
-    async fn test_select_web_page() {
-        let db = setup(None).await;
-        let web_page_res = db
-            .select_web_page(vec!["web:nobc02c8ffyol3kqbsln"])
-            .await
-            .unwrap();
-        println!("web_page_res: {:?}", web_page_res);
-    }
-
-    #[test(tokio::test)]
-    async fn test_select_document() {
-        let db = setup(None).await;
-        let document_res = db
-            .select_document(vec!["document:6dr6glzpf7ixefh7vjks"])
-            .await
-            .unwrap();
-        println!("document_res: {:?}", document_res);
-    }
-
-    #[test(tokio::test)]
-    async fn test_backtrace_by_ids() {
-        let _db = setup(None).await;
-        unimplemented!("TODO!")
-    }
 
     #[test(tokio::test)]
     async fn test_full_text_search_with_highlight() {
